+++ NEWS.md
@@ -11,6 +11,8 @@ Version 0.16.3 - dev
 - CHANGE: move `MText.set_color()` to `MTextEditor.change_color()`
 - CHANGE: move `MText.append_stacked_text()` to `MTextEditor.stacked_text()`
 - BUGFIX: DXF export of GROUP checks for deleted entities
+- BUGFIX: `DIMENSION` transformation also transform the content of the 
+  associated anonymous geometry block content
 
 Version 0.16.2 - 2021-04-21
 ---------------------------